feature. 修改分页中字段数据类型为long类型, 保持与Mybatis中IPage字段统一

+++ hc-base/src/main/java/pro/haichuang/framework/base/constant/PageConstant.java
@@ -15,17 +15,17 @@ public class PageConstant {
     /**
      * 默认页码
      */
-    public static final int DEFAULT_PAGE_NO = 1;
+    public static final long DEFAULT_PAGE_NO = 1;
 
     /**
      * 默认每页展示数量
      */
-    public static final int DEFAULT_PAGE_SIZE = 10;
+    public static final long DEFAULT_PAGE_SIZE = 10;
 
     /**
      * 默认最大每页数量
      */
-    public static final int DEFAULT_MAX_PAGE_SIZE = 1000;
+    public static final long DEFAULT_MAX_PAGE_SIZE = 1000;
 
     /**
      * 默认总数
